+++ lang/es-mx.js
@@ -3,7 +3,9 @@
  * For licensing, see LICENSE.md or http://ckeditor.com/license
  */
 
- // MISSING
+/**
+* @fileOverview 
+*/
 
 /**#@+
    @type String
@@ -109,6 +111,8 @@ CKEDITOR.lang[ 'es-mx' ] = {
 		},
 
 		// Prepended to ARIA labels with shortcuts.
-		keyboardShortcut: 'Atajo de teclado'
+		keyboardShortcut: 'Atajo de teclado',
+
+		optionDefault: 'Default' // MISSING
 	}
 };