+++ lang/hu.js
@@ -112,6 +112,8 @@ CKEDITOR.lang[ 'hu' ] = {
 		},
 
 		// Prepended to ARIA labels with shortcuts.
-		keyboardShortcut: 'Gyorsbillentyű'
+		keyboardShortcut: 'Gyorsbillentyű',
+
+		optionDefault: 'Default' // MISSING
 	}
 };