+++ lang/id.js
@@ -111,6 +111,8 @@ CKEDITOR.lang[ 'id' ] = {
 		},
 
 		// Prepended to ARIA labels with shortcuts.
-		keyboardShortcut: 'Pintasan Keyboard'
+		keyboardShortcut: 'Pintasan Keyboard',
+
+		optionDefault: 'Default' // MISSING
 	}
 };