+++ lang/sk.js
@@ -112,6 +112,8 @@ CKEDITOR.lang[ 'sk' ] = {
 		},
 
 		// Prepended to ARIA labels with shortcuts.
-		keyboardShortcut: 'Klávesová skratka'
+		keyboardShortcut: 'Klávesová skratka',
+
+		optionDefault: 'Default' // MISSING
 	}
 };