+++ lang/sv.js
@@ -111,6 +111,8 @@ CKEDITOR.lang[ 'sv' ] = {
 		},
 
 		// Prepended to ARIA labels with shortcuts.
-		keyboardShortcut: 'Kortkommando'
+		keyboardShortcut: 'Kortkommando',
+
+		optionDefault: 'Default' // MISSING
 	}
 };